rewrite aop annotation and refactoring
