+++ src/Arthurh/Sphring/SphringBoot/SphringBootAnnotation.php
@@ -19,10 +19,10 @@ use Arthurh\Sphring\EventDispatcher\Listener\AnnotationMethodCallAfterListener;
 use Arthurh\Sphring\EventDispatcher\Listener\AnnotationMethodCallBeforeListener;
 use Arthurh\Sphring\EventDispatcher\Listener\AnnotationMethodListener;
 use Arthurh\Sphring\EventDispatcher\SphringEventDispatcher;
-use Arthurh\Sphring\Model\Annotation\AfterCallAnnotation;
 use Arthurh\Sphring\Model\Annotation\AfterLoadMethodOnSphringEventAnnotation;
+use Arthurh\Sphring\Model\Annotation\AopAnnotation\AfterCallAnnotation;
+use Arthurh\Sphring\Model\Annotation\AopAnnotation\BeforeCallAnnotation;
 use Arthurh\Sphring\Model\Annotation\AutoWireAnnotation;
-use Arthurh\Sphring\Model\Annotation\BeforeCallAnnotation;
 use Arthurh\Sphring\Model\Annotation\BeforeLoadMethodOnSphringEventAnnotation;
 use Arthurh\Sphring\Model\Annotation\BeforeStartMethodOnSphringEventAnnotation;
 use Arthurh\Sphring\Model\Annotation\ComposerLockFileAnnotation;
